requirements.txt
# Context-CrackNet Requirements

# Core dependencies
torch>=1.9.0
torchvision>=0.10.0

# Segmentation models for baselines
segmentation-models-pytorch>=0.3.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0
albumentations>=1.0.0

# Scientific computing
numpy>=1.19.0
scipy>=1.5.0

# Configuration
PyYAML>=5.4.0

# Progress bars
tqdm>=4.60.0

# Optional: For development
# matplotlib>=3.3.0
# tensorboard>=2.4.0
